Move type/trait aliases to futures-util
