+++ futures-util/src/compat/compat03as01.rs
@@ -6,7 +6,7 @@ use futures_01::{
 use futures_01::{
     AsyncSink as AsyncSink01, Sink as Sink01, StartSend as StartSend01,
 };
-use futures_core::{
+use crate::{
     task::{RawWaker, RawWakerVTable},
     future::TryFuture as TryFuture03,
     stream::TryStream as TryStream03,
@@ -27,8 +27,8 @@ use std::{
     task::Context,
 };
 
-/// Converts a futures 0.3 [`TryFuture`](futures_core::future::TryFuture) or
-/// [`TryStream`](futures_core::stream::TryStream) into a futures 0.1
+/// Converts a futures 0.3 [`TryFuture`](crate::future::TryFuture) or
+/// [`TryStream`](crate::stream::TryStream) into a futures 0.1
 /// [`Future`](futures_01::future::Future) or
 /// [`Stream`](futures_01::stream::Stream).
 #[derive(Debug, Clone, Copy)]
