+++ futures-util/src/future/try_future/try_flatten_err.rs
@@ -1,5 +1,5 @@
+use crate::future::{FusedFuture, Future, TryFuture};
 use core::pin::Pin;
-use futures_core::future::{FusedFuture, Future, TryFuture};
 use futures_core::ready;
 use futures_core::task::{Context, Poll};
 use pin_project_lite::pin_project;
