+++ futures-util/src/sink/send_all.rs
@@ -1,9 +1,8 @@
-use crate::stream::{StreamExt, TryStreamExt, Fuse};
+use crate::future::Future;
+use crate::stream::{Fuse, Stream, StreamExt, TryStream, TryStreamExt};
 use core::fmt;
 use core::pin::Pin;
-use futures_core::future::Future;
 use futures_core::ready;
-use futures_core::stream::{TryStream, Stream};
 use futures_core::task::{Context, Poll};
 use futures_sink::Sink;
