+++ futures-util/src/stream/try_stream/into_async_read.rs
@@ -1,7 +1,6 @@
-use crate::stream::TryStreamExt;
+use crate::stream::{TryStream, TryStreamExt};
 use core::pin::Pin;
 use futures_core::ready;
-use futures_core::stream::TryStream;
 use futures_core::task::{Context, Poll};
 use futures_io::{AsyncRead, AsyncWrite, AsyncBufRead};
 use std::cmp;
