+++ futures-util/src/stream/try_stream/into_stream.rs
@@ -1,5 +1,5 @@
+use crate::stream::{FusedStream, Stream, TryStream};
 use core::pin::Pin;
-use futures_core::stream::{FusedStream, Stream, TryStream};
 use futures_core::task::{Context, Poll};
 #[cfg(feature = "sink")]
 use futures_sink::Sink;
