+++ futures-util/src/stream/try_stream/or_else.rs
@@ -1,8 +1,8 @@
+use crate::future::TryFuture;
+use crate::stream::{Stream, TryStream, FusedStream};
 use core::fmt;
 use core::pin::Pin;
-use futures_core::future::TryFuture;
 use futures_core::ready;
-use futures_core::stream::{Stream, TryStream, FusedStream};
 use futures_core::task::{Context, Poll};
 #[cfg(feature = "sink")]
 use futures_sink::Sink;
