+++ futures-util/src/stream/try_stream/try_collect.rs
@@ -1,8 +1,8 @@
+use crate::future::{FusedFuture, Future};
+use crate::stream::{FusedStream, TryStream};
 use core::mem;
 use core::pin::Pin;
-use futures_core::future::{FusedFuture, Future};
 use futures_core::ready;
-use futures_core::stream::{FusedStream, TryStream};
 use futures_core::task::{Context, Poll};
 use pin_project_lite::pin_project;
