+++ futures-util/src/stream/try_stream/try_concat.rs
@@ -1,7 +1,7 @@
+use crate::stream::TryStream;
+use crate::future::Future;
 use core::pin::Pin;
-use futures_core::future::Future;
 use futures_core::ready;
-use futures_core::stream::TryStream;
 use futures_core::task::{Context, Poll};
 use pin_project_lite::pin_project;
