+++ futures-util/src/stream/try_stream/try_flatten.rs
@@ -1,6 +1,6 @@
+use crate::stream::{FusedStream, Stream, TryStream};
 use core::pin::Pin;
 use futures_core::ready;
-use futures_core::stream::{FusedStream, Stream, TryStream};
 use futures_core::task::{Context, Poll};
 #[cfg(feature = "sink")]
 use futures_sink::Sink;
