+++ futures-util/src/stream/try_stream/try_fold.rs
@@ -1,8 +1,8 @@
+use crate::future::{FusedFuture, Future, TryFuture};
+use crate::stream::TryStream;
 use core::fmt;
 use core::pin::Pin;
-use futures_core::future::{FusedFuture, Future, TryFuture};
 use futures_core::ready;
-use futures_core::stream::TryStream;
 use futures_core::task::{Context, Poll};
 use pin_project_lite::pin_project;
